# docs/swagger/swagger.yaml
definitions:
...
  ConnectSessionResponse:
    properties:
      message:
        example: Session connection initiated successfully
        type: string
      qrCode:
        example: 2@abc123...
        type: string
      qrCodeImage:
        example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  CreateSessionRequest:
    properties:
      name:
        example: my-session
        maxLength: 50
        minLength: 3
        type: string
      proxyConfig:
        $ref: '#/definitions/ProxyConfig'
      qrCode:
        example: false
        type: boolean
    required:
    - name
    type: object
  CreateSessionResponse:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1b2e424c-a2a0-41a4-b992-15b7ec06b9bc
        type: string
      isConnected:
        example: false
        type: boolean
      name:
        example: my-session
        type: string
      proxyConfig:
        $ref: '#/definitions/ProxyConfig'
      qrCode:
        example: 2@abc123...
        type: string
      qrCodeImage:
        example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
        type: string
    type: object
  DeviceInfoResponse:
    properties:
      appVersion:
        example: 2.21.4.18
        type: string
      deviceModel:
        example: Samsung Galaxy S21
        type: string
      osVersion:
        example: "11"
        type: string
      platform:
        example: android
        type: string
    type: object
...
  ErrorResponse:
    properties:
      code:
        example: VALIDATION_ERROR
        type: string
      details: {}
      error:
        example: Invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  ListSessionsResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      sessions:
        items:
          $ref: '#/definitions/SessionInfoResponse'
        type: array
      total:
        example: 10
        type: integer
    type: object
...
  PairPhoneRequest:
    properties:
      phoneNumber:
        example: "+5511999999999"
        type: string
    required:
    - phoneNumber
    type: object
...
  ProxyResponse:
    properties:
      proxyConfig:
        $ref: '#/definitions/ProxyConfig'
    type: object
  QRCodeResponse:
    properties:
      expiresAt:
        example: "2024-01-01T00:01:00Z"
        type: string
      qrCode:
        example: 2@abc123def456...
        type: string
      qrCodeImage:
        example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
        type: string
      timeoutSeconds:
        example: 60
        type: integer
    type: object
...
  SessionInfoResponse:
    properties:
      deviceInfo:
        $ref: '#/definitions/DeviceInfoResponse'
      session:
        $ref: '#/definitions/SessionResponse'
    type: object
...
  SessionStatsResponse:
    properties:
      connected:
        example: 3
        type: integer
      offline:
        example: 7
        type: integer
      total:
        example: 10
        type: integer
    type: object
  SetProxyRequest:
    properties:
      proxyConfig:
        $ref: '#/definitions/ProxyConfig'
    required:
    - proxyConfig
    type: object
  SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /sessions/{sessionId}/connect:
    post:
      description: Connect a WhatsApp session to start receiving messages. Automatically
        returns QR code (both string and base64 image) if device needs to be paired.
        If session is already connected, returns confirmation message.
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session connection initiated successfully with QR code if needed,
            or confirmation if already connected
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/ConnectSessionResponse'
              type: object
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect session
      tags:
      - Sessions
...
  /sessions/{sessionId}/delete:
    delete:
      description: Delete a WhatsApp session and all associated data
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete session
      tags:
      - Sessions
  /sessions/{sessionId}/disconnect:
    post:
      description: Disconnect from WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session disconnected successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disconnect session
      tags:
      - Sessions
...
  /sessions/{sessionId}/info:
    get:
      description: Get detailed information about a specific WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session information retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/SessionInfoResponse'
              type: object
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get session information
      tags:
      - Sessions
  /sessions/{sessionId}/logout:
    post:
      description: Logout from WhatsApp session and disconnect
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session logout successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout session
      tags:
      - Sessions
...
  /sessions/{sessionId}/pair:
    post:
      consumes:
      - application/json
      description: Pair WhatsApp session with phone number
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone pairing request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PairPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Phone pairing initiated successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Pair phone number
      tags:
      - Sessions
  /sessions/{sessionId}/proxy:
    get:
      description: Get proxy configuration for a WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proxy configuration retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/ProxyResponse'
              type: object
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get proxy
      tags:
      - Sessions
  /sessions/{sessionId}/proxy/set:
    post:
      consumes:
      - application/json
      description: Configure proxy settings for a WhatsApp session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Proxy configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/SetProxyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Proxy configured successfully
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Set proxy
      tags:
      - Sessions
  /sessions/{sessionId}/qr:
    get:
      description: Get QR code for WhatsApp session pairing. Returns both raw QR code
        string and base64 image.
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: QR code generated successfully with base64 image
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/QRCodeResponse'
              type: object
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get QR code
      tags:
      - Sessions
  /sessions/{sessionId}/qr/generate:
    post:
      description: Generate a new QR code for WhatsApp session pairing
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: QR code generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/QRCodeResponse'
              type: object
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Generate QR code
      tags:
      - Sessions
...
  /sessions/create:
    post:
      consumes:
      - application/json
      description: Create a new WhatsApp session with optional proxy configuration.
        If qrCode is true, returns QR code immediately for connection.
      parameters:
      - description: Session creation request with optional qrCode flag
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Session created successfully. If qrCode was true, includes
            QR code data.
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/CreateSessionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Session already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new session
      tags:
      - Sessions
  /sessions/list:
    get:
      consumes:
      - application/json
      description: Get a list of all WhatsApp sessions with optional filtering
      parameters:
      - description: Filter by connection status
        in: query
        name: isConnected
        type: boolean
      - description: Filter by device JID
        in: query
        name: deviceJid
        type: string
      - description: 'Number of sessions to return (default: 20)'
        in: query
        name: limit
        type: integer
      - description: 'Number of sessions to skip (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Sessions retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/ListSessionsResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List sessions
      tags:
      - Sessions
  /sessions/stats:
    get:
      description: Get statistics about all sessions
      produces:
      - application/json
      responses:
        "200":
          description: Session statistics retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/SessionStatsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get session statistics
      tags:
      - Sessions
swagger: "2.0"